Refactor errors
