+++ packages/gbnf/javascript/src/grammar-graph/graph.ts
@@ -18,7 +18,7 @@ import {
   isRuleRef,
 } from './type-guards.js';
 import { isPointInRange, } from "../utils/is-point-in-range.js";
-import { InputParseError, } from "../utils/errors/errors.js";
+import { InputParseError, } from "../utils/errors/input-parse-error.js";
 import { RuleRef, } from "./rule-ref.js";
 import { getInputAsCodePoints, } from "./get-input-as-code-points.js";
