+++ packages/gbnf/javascript/src/rules-builder/parse-char.ts
@@ -1,4 +1,4 @@
-import { GrammarParseError, } from "../utils/errors/errors.js";
+import { GrammarParseError, } from "../utils/errors/input-parse-error.js";
 
 export const parseChar = (src: string, pos: number): [number, number] => {
   if (src[pos] === '\\') {
