+++ packages/gbnf/javascript/src/rules-builder/rules-builder.ts
@@ -1,4 +1,4 @@
-import { GrammarParseError, } from "../utils/errors/errors.js";
+import { GrammarParseError, } from "../utils/errors/input-parse-error.js";
 import { isWordChar, } from "./is-word-char.js";
 import { parseChar, } from "./parse-char.js";
 import { parseName, } from "./parse-name.js";
