+++ packages/gbnf/javascript/src/umd.ts
@@ -13,7 +13,7 @@ import { ParseState, } from './grammar-graph/parse-state.js';
 import {
   InputParseError,
   GrammarParseError,
-} from './utils/errors/errors.js';
+} from './utils/errors/input-parse-error.js';
 
 module.exports = GBNF;
 module.exports.RuleType = InternalRuleType;
