+++ packages/gbnf/javascript/src/utils/errors/get-input-as-string.ts
@@ -1,0 +1,11 @@
+import type {
+  ValidInput,
+} from "../../grammar-graph/types.js";
+
+export const getInputAsString = (src: ValidInput) => {
+  if (typeof src === 'string') {
+    return src;
+  }
+  const codePoints = Array.isArray(src) ? src : [src,];
+  return codePoints.map(cp => String.fromCodePoint(cp)).join('');
+};
